+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+**Unreleased:**
+---------------
+
+**Add:**
+
+* Add print_dict and print_ini methods.
+
 
 v0.2.0 (2020-03-21)
 --------------------