+++ CHANGELOG.rst
@@ -15,6 +15,10 @@ Changelog
 
 * Add print_dict and print_ini methods.
 
+* Add DictOptionBuilder.
+
+  (Old OptionPaser has renamed to FiniOptionBuilder).
+
 
 v0.2.0 (2020-03-21)
 --------------------